Make it possible to set any country package
